removed RT from centroid struct, removed const which could break move semantics, changed readCSV function to minimise type conversions There is still an error in the transfer function which leads to more scans being created than exist in the data.

+++ include/qalgorithms_measurement_data_lcms.h
@@ -42,7 +42,7 @@ namespace q
              * @param variableTypes The variable types. Required variable types: DataField::SCANNUMBER,
              *                      DataField::RETENTIONTIME, DataField::MZ, and DataField::INTENSITY
              */
-            bool readCSV(std::string filename, int rowStart, int rowEnd,
+            bool readCSV(std::string filename, size_t rowStart, int rowEnd,
                          int colStart, int colEnd, char separator,
                          std::vector<DataType::DataField> variableTypes);
